+++ fibjs/src/BufferedStream.cpp
@@ -210,8 +210,8 @@ result_t BufferedStream::onwrite(v8::Handle<v8::Function> func)
 	return on("write", func);
 }
 
-result_t BufferedStream::copyTo(obj_ptr<Stream_base>& stm, int32_t bytes,
-		int32_t& retVal, exlib::AsyncEvent* ac)
+result_t BufferedStream::copyTo(obj_ptr<Stream_base>& stm, int64_t bytes,
+		int64_t& retVal, exlib::AsyncEvent* ac)
 {
 	if (!ac)
 		return CALL_E_NOSYNC;
@@ -219,7 +219,7 @@ result_t BufferedStream::copyTo(obj_ptr<Stream_base>& stm, int32_t bytes,
 	return copyStream(this, stm, bytes, retVal, ac);
 }
 
-result_t BufferedStream::asyncCopyTo(obj_ptr<Stream_base>& stm, int32_t bytes)
+result_t BufferedStream::asyncCopyTo(obj_ptr<Stream_base>& stm, int64_t bytes)
 {
 	acb_copyTo(s_acPool, stm, bytes);
 	return 0;
